+++ packages/instrumentation-typeorm/src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './get-func-param-names';
+export * from './suppressTracing';